% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_proc_format.R
\name{read_proc_format}
\alias{read_proc_format}
\title{Read Proc Format File}
\usage{
read_proc_format(file)
}
\arguments{
\item{file}{Path to the SAS formats file}
}
\description{
Read Proc Format File
}
\examples{
\dontrun{
read_proc_format("pu_ssocs18_format.txt")
}

}
\seealso{
Other SAS helper functions: 
\code{\link{read_proc_format_vars}()},
\code{\link{read_ssocs}()}
}
\concept{SAS helper functions}
