% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_proc_format.R
\name{read_proc_format_vars}
\alias{read_proc_format_vars}
\title{Read Proc Fmt File}
\usage{
read_proc_format_vars(file)
}
\arguments{
\item{file}{Path to the SAS fmts file}
}
\description{
Read Proc Fmt File
}
\examples{
\dontrun{
read_proc_format_vars("pu_ssocs18_format.txt")
}

}
\seealso{
Other SAS helper functions: 
\code{\link{read_proc_format}()},
\code{\link{read_ssocs}()}
}
\concept{SAS helper functions}
