% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ssocs.R
\name{read_ssocs}
\alias{read_ssocs}
\title{Read SAS file and value labels according to PROC FORMAT}
\usage{
read_ssocs(
  datafile,
  formatfile,
  fmtfile = NULL,
  as_factor = TRUE,
  ordered = TRUE
)
}
\arguments{
\item{datafile}{Path to the `sas7bdat` file to be read by [haven::read_sas]}

\item{formatfile}{Path to the SAS formats file to be read by [read_proc_format]}

\item{fmtfile}{Path to the SAS fmts file to be read by [read_proc_format_vars]}

\item{as_factor}{If TRUE, onvert labelled variables to factor with [haven::as_factor]}

\item{ordered}{If TRUE and as_factor is TRUE, convert to ordered factor}
}
\value{
A tibble, data frame variant with nice defaults.

  Variable labels are stored in the "label" attribute of each variable.
  It is not printed on the console, but the RStudio viewer will show it.

  The vectors may be a labelled vector that may be coerced into a
  factor or ordered factor.
}
\description{
Reads a SAS `sas7bdat` file using [haven::read_sas] and then applies the
labels given in the formats file (containing `PROC FORMAT` statements).
The labels use [labelled::labelled] and may be coerced to factors or ordered
factors if desired.
}
\examples{
\dontrun{
read_ssocs("pu_ssocs18.sas7bdat", "pu_ssocs18_format.txt")
}

}
\seealso{
Other SAS helper functions: 
\code{\link{read_proc_format_vars}()},
\code{\link{read_proc_format}()}
}
\concept{SAS helper functions}
